refactor: BookDocument를 document  패키지로 이동
